+++ Exercism/c/hamming/src/hamming.h
@@ -1,0 +1,6 @@
+#ifndef HAMMING_H
+#define HAMMING_H
+
+int compute(const char* one, const char* two);
+
+#endif