+++ Exercism/c/raindrops/src/raindrops.h
@@ -1,0 +1,8 @@
+#ifndef RAINDROPS_H
+#define RAINDROPS_H
+
+#include <stdlib.h>
+
+char* convert(char buffer[], size_t size, int num);
+
+#endif