+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 #### 23.06.2017
 
+* [C] Hamming
 * [C] Gigasecond (Exercism)
 
 #### 22.06.2017